generate valid package for any role arguments, including refs
This means that the role arguments can be anything, and interpreting
them will be left to the role library.  For example, Moose allows
passing a hashref after the role name to with, which can include version
checks or parameterized role configuration.

For refs, this will end up including the stringified ref.  This isn't
particularly friendly, but is pretty much the best we can do.