load roles before validating them against Role::Tiny
When detecting the library to use based on the base class, we can just
check for which libraries are already loaded, since the base class
already has to have been loaded.  When detecting based on roles, the
roles may not be loaded yet, so we need to load them first.