add with::roles::anon function to create subclasses that clean up their packages when the object is destroyed
Normally, the packages created by this module are left in place even
when all of the objects are removed.  This is usually the correct
choice, because the base class and roles you use will very likely be the
same.  However, you may be dynamically be creating roles, which would
mean the generated packages would never be used again.  The
with::roles::anon function only works on objects, but when the object is
destroyed, the relevant package is deleted.